backend/requirements.txt
Flask
mysql-connector-python>=8.0.11
Flask-CORS
